third_party/blink/web_tests/external/wpt/mixed-content/gen/srcdoc-inherit.http-rp/opt-in/worklet-animation.https-expected.txt

This is a testharness.js-based test.
[FAIL] Mixed-Content: Expects allowed for worklet-animation to same-https origin and keep-scheme redirection from https context.
  assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
[FAIL] Mixed-Content: Expects allowed for worklet-animation to same-https origin and no-redirect redirection from https context.
  assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
Harness: the test ran to completion.
